# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clothing e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: size
          in: query
          schema:
            type: string
        - name: color
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # Cart
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user's wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # Orders
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/addresses:
    get:
      operationId: getUserAddresses
      tags: [users]
      summary: Get user addresses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addUserAddress
      tags: [users]
      summary: Add address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, staff]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, staff]

    # Category
...
    Product:
      type: object
      required: [id, title, price, category, stock, images, sizes, colors, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        category:
          type: string
        categoryId:
          type: ["integer", "null"]
        images:
          type: array
          items:
            type: string
        modelUrl:
          type: ["string", "null"]
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [title, price, category, stock]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        category:
          type: string
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        modelUrl:
          type: string
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        category:
          type: string
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        modelUrl:
          type: string
        stock:
          type: integer
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Cart
...
    WishlistItem:
      type: object
      required: [id, productId, product, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
          format: date-time
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer

    # Order
...
    Address:
      type: object
      required: [id, userId, line1, city, state, country, zip]
      properties:
        id:
          type: integer
        userId:
          type: integer
        label:
          type: ["string", "null"]
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zip:
          type: string
        isDefault:
          type: boolean
    AddressInput:
      type: object
      required: [line1, city, state, country, zip]
      properties:
        label:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        zip:
          type: string
        isDefault:
          type: boolean

    # Admin Stats
...