# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clothing e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  compareAtPrice?: number;
  category: string;
  categoryId?: number;
  images?: string[];
  modelUrl?: string;
  stock: number;
  sizes?: string[];
  colors?: string[];
  tags?: string[];
  featured?: boolean;
}